requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-autocomplete-light==3.9.7
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-sql-explorer==3.2.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
